requirements.txt
pandas
numpy
psycopg2-binary
pymongo
python-dotenv
PyYAML